requirements.txt
Pillow>=10.0
streamlit>=1.32
pandas>=2.2
openpyxl>=3.1
python-docx>=1.2
